feat(provider): :zap: bailey add send file video audio
